Backend/requirements.txt
# Core framework
fastapi==0.111.0         # stable with Pydantic v2
uvicorn[standard]==0.30.1

# YOLOv12 (Ultralytics family)
ultralytics==8.3.0       # supports YOLOv8 → YOLOv12
torch==2.2.2
torchvision==0.17.2

# Image / Video Processing
pillow==10.3.0
opencv-python==4.9.0.80

# Database (PostgreSQL + ORM)
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Data validation & settings
pydantic==2.8.2
pydantic-settings==2.3.4

# API client (Gemini API + requests)
google-generativeai==0.7.2
httpx==0.27.0

# Text-to-Speech (TTS)
gTTS==2.5.1
pyttsx3==2.90

# Utils
python-multipart==0.0.9    # For file uploads
python-dotenv==1.0.1       # For .env handling
loguru==0.7.2              # For structured logging